Created methods and such to make the chatbot ask for an input then go on a loop with the user just asking questions and ends.
Chatbot will quit when the word "quit" (regardless of case) is entered but this will not pass the ChatbotTest for the quitter. Also created a string array with 10 questions that it randomizes through.

NOTE: With premission a certain part was changed in the Test so that it will test correctly. (changed "floofer" to "FLOOFER"