Made the shoppingListChecker be able to pass the testShoppingListChecker.
SELF-ADVICE: you do not to check each item individually using && and the exact name for the item. I had to use shoppingItem because all of the things being checked in the test are in shoppingItem.

So essentially, you just need to see if an item in general (no matter what is in the test) is there in the List then it will pass. No for loops needed either.