Added ScrollPane components and changed code accordingly.
Also cahnged the this.add(chatArea) into this.add(chatScrollPane) as well as other areas that have chatArea in the setupLayout() because the program will now be looking at the Scoll Pane.

Also fixed an error with an array size for followup being too large because 4 means there are actually 5 places due to how indexes work.