Added the jar file for the Twitter API to the projecct and added the import and members needed for the controller.
The reason the jar file was placed inside of the project was because if you want to use the program from another computer through gitHub you can just have the downloaded file from there. The new CTECTwitter initialization for "myTwitter" also needs to be BEFORE the view because it will be doing things as well before things are displayed.

CTECTwitter also CAN'T be called "Twitter" because there is a twitter class so names would end up matching which is not allowed unless you put in the entire package specification which is a pain.