Fixed up some code to make the tweet button actually work and sent out my first tweet!
The tweet was sent out to my class twitter account that I made. It will be deleted once the project is done but so far it works because the properties folder is altered so it has my personal keys. That is why the file is inaccessable or at least should be through gitHub because there would be a lot of security issues.