Starting the Query assignment with teacher given code.
We will have to be able to change and add to the code to the point where we can use our program to search a query. The twitter4j website in the 4.0.4 section has all the classes to research on. Look at the other code we've done also for help.